+++ app/src/main/java/edu/calvin/cs262/pilot/knightrank/FollowPUTLoader.java
@@ -1,0 +1,4 @@
+package edu.calvin.cs262.pilot.knightrank;
+
+public class FollowPUTLoader {
+}
