+++ app/src/main/java/edu/calvin/cs262/pilot/knightrank/TestPOSTPUTSportBackEnd.java
@@ -28,30 +28,18 @@ import android.widget.TextView;
 
 import java.util.Objects;
 
-public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderManager.LoaderCallbacks<String>{
+public class TestPOSTSportBackEnd extends AppCompatActivity implements LoaderManager.LoaderCallbacks<String> {
 
     // Private class members.
-    private static final String LOG_TAG = TestPOSTMatchBackEnd.class.getSimpleName();
+    private static final String LOG_TAG = TestPOSTSportBackEnd.class.getSimpleName();
 
-    private EditText editTextViewMatchSportID;
-    private EditText editTextViewMatchPlayerOneID;
-    private EditText editTextViewMatchPlayerTwoID;
-    private EditText editTextViewMatchPlayerOneScore;
-    private EditText editTextViewMatchPlayerTwoScore;
-    private EditText editTextViewMatchWinner;
-    private EditText editTextViewMatchTimeStamp;
-    private EditText editTextViewMatchVerification;
+    private EditText editTextViewSportName;
+    private EditText editTextViewSportType;
     private TextView textViewNetworkStatus;
     private TextView textViewRequestStatus;
 
-    private String matchSportIDString;
-    private String matchPlayerOneIDString;
-    private String matchPlayerTwoIDString;
-    private String matchPlayerOneScoreString;
-    private String matchPlayerTwoScoreString;
-    private String matchWinnerString;
-    private String matchTimeStampString;
-    private String matchVerificationString;
+    private String sportNameString;
+    private String sportTypeString;
 
     // Share preferences file (custom)
     private SharedPreferences mPreferences;
@@ -64,7 +52,7 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_test_postmatch_back_end);
+        setContentView(R.layout.activity_test_postsport_back_end);
 
         // my_child_toolbar is defined in the layout file
         Toolbar myToolbar = (Toolbar) findViewById(R.id.my_toolbar);
@@ -86,25 +74,19 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
         String syncFreq = mPreferencesDefault.getString(SettingsActivity.KEY_SYNC_FREQUENCY, "-1");
 
         // Find components.
-        editTextViewMatchSportID = findViewById(R.id.input_match_sport_id);
-        editTextViewMatchPlayerOneID = findViewById(R.id.input_match_player_one_id);
-        editTextViewMatchPlayerTwoID = findViewById(R.id.input_match_player_two_id);
-        editTextViewMatchPlayerOneScore = findViewById(R.id.input_match_player_one_score);
-        editTextViewMatchPlayerTwoScore = findViewById(R.id.input_match_player_two_score);
-        editTextViewMatchWinner = findViewById(R.id.input_match_winner);
-        editTextViewMatchTimeStamp = findViewById(R.id.input_match_time);
-        editTextViewMatchVerification = findViewById(R.id.input_match_verified);
+        editTextViewSportName = findViewById(R.id.input_sport_name);
+        editTextViewSportType = findViewById(R.id.input_sport_type);
         textViewNetworkStatus = findViewById(R.id.network_connected);
         textViewRequestStatus = findViewById(R.id.request_status);
 
         // Reconnect to the loader if one exists already, upon device config change.
-        if(getSupportLoaderManager().getLoader(0)!=null){
-            getSupportLoaderManager().initLoader(0,null,this);
+        if (getSupportLoaderManager().getLoader(0) != null) {
+            getSupportLoaderManager().initLoader(0, null, this);
         }
 
         // Change the background color to what was selected in color picker.
         // Note: Change color by using findViewById and ID of the UI element you wish to change.
-        RelativeLayout thisLayout = findViewById(R.id.activity_test_match_post_back_end_root_layout);
+        RelativeLayout thisLayout = findViewById(R.id.activity_test_sport_post_back_end_root_layout);
         thisLayout.setBackgroundColor(mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.YELLOW));
 
         int value = mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.BLACK);
@@ -114,7 +96,15 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
         // Change the toolbar color to what was selected in color picker.
         getSupportActionBar().setBackgroundDrawable(new ColorDrawable(toolbarColor));
 
-        Log.e(LOG_TAG,"Value of color is: " + value);
+        Log.e(LOG_TAG, "Value of color is: " + value);
+    }
+
+    /**
+     * Method begins Google Cloud endpoint put process upon button click.
+     *
+     * @param view view component
+     */
+    public void put(View view) {
     }
 
     /**
@@ -125,38 +115,14 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
     public void post(View view) {
 
         // Get the sport input strings.
-        matchSportIDString = editTextViewMatchSportID.getText().toString();
-        matchPlayerOneIDString = editTextViewMatchPlayerOneID.getText().toString();
-        matchPlayerTwoIDString = editTextViewMatchPlayerTwoID.getText().toString();
-        matchPlayerOneScoreString = editTextViewMatchPlayerOneScore.getText().toString();
-        matchPlayerTwoScoreString = editTextViewMatchPlayerTwoScore.getText().toString();
-        matchWinnerString = editTextViewMatchWinner.getText().toString();
-        matchTimeStampString = editTextViewMatchTimeStamp.getText().toString();
-        matchVerificationString = editTextViewMatchVerification.getText().toString();
+        sportNameString = editTextViewSportName.getText().toString();
+        sportTypeString = editTextViewSportType.getText().toString();
 
-        if(matchSportIDString.length() == 0){
-            matchSportIDString = "1";
+        if (sportNameString.length() == 0) {
+            sportNameString = "default sport name";
         }
-        if(matchPlayerOneIDString.length() == 0){
-            matchPlayerOneIDString = "1";
-        }
-        if(matchPlayerTwoIDString.length() == 0){
-            matchPlayerTwoIDString = "2";
-        }
-        if(matchPlayerOneScoreString.length() == 0){
-            matchPlayerOneScoreString = "10";
-        }
-        if(matchPlayerTwoScoreString.length() == 0){
-            matchPlayerTwoScoreString = "20";
-        }
-        if(matchWinnerString.length() == 0){
-            matchWinnerString = "2";
-        }
-        if(matchTimeStampString.length() == 0){
-            matchTimeStampString = "2018-11-03 00:53:57.048546";
-        }
-        if(matchVerificationString.length() == 0){
-            matchVerificationString = "false";
+        if (sportTypeString.length() == 0) {
+            sportTypeString = "default sport type";
         }
 
         // Close keyboard after hitting search query button.
@@ -181,21 +147,15 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
         if (networkInfo != null && networkInfo.isConnected()) {
 
             // show "Connected" & type of network "WIFI or MOBILE"
-            textViewNetworkStatus.setText("Connected "+networkInfo.getTypeName());
+            textViewNetworkStatus.setText("Connected " + networkInfo.getTypeName());
             // change background color to red
             textViewNetworkStatus.setBackgroundColor(0xFF7CCC26);
 
             // Refactored to user AsyncTaskLoader via PlayerGETLoader.java
             Bundle postBundle = new Bundle();
-            postBundle.putString("match_sport_id", matchSportIDString);
-            postBundle.putString("match_player_one_id", matchPlayerOneIDString);
-            postBundle.putString("match_player_two_id", matchPlayerTwoIDString);
-            postBundle.putString("match_player_one_score", matchPlayerOneScoreString);
-            postBundle.putString("match_player_two_score", matchPlayerTwoScoreString);
-            postBundle.putString("match_winner", matchWinnerString);
-            postBundle.putString("match_timestamp", matchTimeStampString);
-            postBundle.putString("match_verified", matchVerificationString);
-            getSupportLoaderManager().restartLoader(0, postBundle,this);
+            postBundle.putString("sport_name", sportNameString);
+            postBundle.putString("sport_type", sportTypeString);
+            getSupportLoaderManager().restartLoader(0, postBundle, this);
 
             // Indicate to user query is in process.
             textViewRequestStatus.setText("");
@@ -252,7 +212,7 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
 
     /**
      * Method called when load is instantiated.
-     *
+     * <p>
      * NOTE: change the Loader.java class if we want to obtain different data.
      *
      * @param id
@@ -262,19 +222,12 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
     @NonNull
     @Override
     public Loader<String> onCreateLoader(int id, @Nullable Bundle bundle) {
-        return new MatchPOSTLoader(this, bundle.getString("match_sport_id"),
-                bundle.getString("match_player_one_id"),
-                bundle.getString("match_player_two_id"),
-                bundle.getString("match_player_one_score"),
-                bundle.getString("match_player_two_score"),
-                bundle.getString("match_winner"),
-                bundle.getString("match_timestamp"),
-                bundle.getString("match_verified"));
+        return new SportPOSTLoader(this, bundle.getString("sport_name"), bundle.getString("sport_type"));
     }
 
     /**
      * Method called when loader task is finished.  Add code to update UI with results.
-     *
+     * <p>
      * NOTE: modify as necessary according to the data we obtained.
      *
      * @param loader loader object.
@@ -285,7 +238,7 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
     public void onLoadFinished(@NonNull Loader<String> loader, String s) {
 
         // Method call to test match GET.
-        TestMatchPOSTBackend(s);
+        TestSportPOSTBackend(s);
     }
 
     /**
@@ -293,7 +246,7 @@ public class TestPOSTMatchBackEnd extends AppCompatActivity implements LoaderMan
      *
      * @param s response message from the RESTful web service.
      */
-    private void TestMatchPOSTBackend(String s) {
+    private void TestSportPOSTBackend(String s) {
 
         // POST the response we get to the TextView.
         textViewRequestStatus.setText("Response from RESTful web service\n");
